+++ pleno/processing/tools/matrix.h
@@ -15,17 +15,17 @@ inline std::array<size_t, 2> index_to_colRow(size_t matrix_width, size_t index)
 
     std::array<size_t, 2> indices;
     indices[1] = index / matrix_width;
-    indices[0] = std::fma(-float(indices[1]), float(matrix_width), float(index));
+    indices[0] = std::fma(-double(indices[1]), double(matrix_width), double(index));
 
     return indices;
 }
 
 inline size_t colRow_to_index(size_t matrix_width, const std::array<size_t, 2>& coords)
 {
-    return std::fma(float(coords[1]), float(matrix_width), float(coords[0]));
+    return std::fma(double(coords[1]), double(matrix_width), double(coords[0]));
 }
 
 inline size_t colRow_to_index(size_t matrix_width, size_t col, size_t row)
 {
-    return std::fma(float(row), float(matrix_width), float(col));
+    return std::fma(double(row), double(matrix_width), double(col));
 }
